Trying to fix memory problems

+++ sql/join_optimizer/estimate_selectivity.cc
@@ -146,7 +146,7 @@ double EstimateSelectivity(THD *thd, Item *condition, string *trace) {
           }
         }
         printf("3\n");
-        selectivity = std::max(-1, std::max(estimatedRowsLeft, estimatedRowsRight)/(estimatedRowsLeft * estimatedRowsRight));
+        selectivity = std::max((double)-1, std::max(estimatedRowsLeft, estimatedRowsRight)/(estimatedRowsLeft * estimatedRowsRight));
         printf("4\n");
       }else if(left->type() == Item::FIELD_ITEM && !(right->type() == Item::FIELD_ITEM)){
         for(Field *field : {down_cast<Item_field *>(left)->field}){
